vision/requirements.txt
python-dotenv
anthropic
google-genai
openai==0.28.1